Isolando metodo New de computers no Repositorys
