+++ Database/DatabaseSetup.cs
@@ -11,7 +11,6 @@ class DatabaseSetup
     {
         _databaseConfig = databaseConfig;
         CreateComputerTable();
-        CreateLabTable();
     }
 
     public void CreateComputerTable()
@@ -30,8 +29,4 @@ class DatabaseSetup
         connection.Close();
     }
 
-    public void CreateLabTable()
-    {
-
-    }
 }
