+++ Models/Computer.cs
@@ -1,6 +1,6 @@
 namespace LabManager.Models;
 
-class Computer
+public class Computer
 {
     public int Id { get; set; }
     public string Ram { get; set; }
